Assorted fixes.
Default to displaying all contacts rather than none as it's confusing
and a litte error prone that the model displays nothing without setting
a filter.

Don't ignore results from refreshed updates, and fix errors not caught
because that wasn't happening.

Use the custom label from QContact if a first/last name isn't available
before falling back to an alternative.

Resolve names from IM accounts correctly in sparql query.

When querying for favorites, test the value not just the existence of
the property.

Fallback to the imAvatar if there is no photo for a contact, and emit
dataChanged if the url for an avatar has changed.