[nemo-qml-plugin-contacts] Adapt to changes in libcontacts for sort/display order
Add the primaryName and secondaryName readonly properties to SeasidePerson,
to be used where display order is required, rather than a formal
separation between first and last names.

Add the primaryName and secondaryName roles to the model, where they
should be used in preference to the existing firstName and lastName
roles, which are now deprecated.