[nemo-qml-plugin-contacts] Add detail accessor/mutator functions
For non-composite details, it is vital to know if the content has been
modified from an existing detail, or if it has been added independently.
Add a mechanism of extracting and re-inserting data that allows the
person to record modifications as such, maintaining the value of any
metadata in the detail record that was not directly changed in an edit
operation.